requirements.txt
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
daphne==4.1.2
websockets==12.0
requests==2.31.0
django-channels==0.7.0
Django==5.1.1
channels==4.1.0
Pyrebase4==4.8.0
pyfcm==2.0.7